Add GitHub Actions workflow for .NET project build
This commit introduces a new workflow configuration file `GH-build-DEV.yml` that automates the build process for a .NET project. The workflow is triggered on pushes to the "main" branch and includes steps for checking out the code, setting up the .NET environment, restoring dependencies, building the project, and running tests. It uses Ubuntu 24.04 and .NET version 9.0.4.